+++ python/GafferSceneUI/__init__.py
@@ -160,6 +160,7 @@ import ShuffleAttributesUI
 import ShufflePrimitiveVariablesUI
 import LocaliseAttributesUI
 import PrimitiveSamplerUI
+import ClosestPointSamplerUI
 
 # then all the PathPreviewWidgets. note that the order
 # of import controls the order of display.